# tools/openapi.json
...
openapi: 3.0.1
info:
  title: XBuilder Server API
  version: "1.0"
...
paths:
  /api/documents/credit-note/create:
    post:
      tags:
      - create
      summary: Create a XML file from the input
      parameters:
      - name: x-openubl-certificatekey
        in: header
        schema:
          type: string
      - name: x-openubl-privatekey
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditNoteInputModel'
      responses:
        "200":
          description: OK
          content:
            text/xml:
              schema:
                type: string
  /api/documents/credit-note/enrich:
    post:
      tags:
      - enrich
      summary: Enriches the input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditNoteInputModel'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditNoteOutputModel'
  /api/documents/debit-note/create:
    post:
      tags:
      - create
      summary: Create a XML file from the input
      parameters:
      - name: x-openubl-certificatekey
        in: header
        schema:
          type: string
      - name: x-openubl-privatekey
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitNoteInputModel'
      responses:
        "200":
          description: OK
          content:
            text/xml:
              schema:
                type: string
  /api/documents/debit-note/enrich:
    post:
      tags:
      - enrich
      summary: Enriches the input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitNoteInputModel'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitNoteOutputModel'
  /api/documents/invoice/create:
    post:
      tags:
      - create
      summary: Create a XML file from the input
      parameters:
      - name: x-openubl-certificatekey
        in: header
        schema:
          type: string
      - name: x-openubl-privatekey
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInputModel'
      responses:
        "200":
          description: OK
          content:
            text/xml:
              schema:
                type: string
  /api/documents/invoice/enrich:
    post:
      tags:
      - enrich
      summary: Enriches the input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInputModel'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceOutputModel'
  /api/documents/summary-document/create:
    post:
      tags:
      - create
      summary: Create a XML file from the input
      parameters:
      - name: x-openubl-certificatekey
        in: header
        schema:
          type: string
      - name: x-openubl-privatekey
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummaryDocumentInputModel'
      responses:
        "200":
          description: OK
          content:
            text/xml:
              schema:
                type: string
  /api/documents/summary-document/enrich:
    post:
      tags:
      - enrich
      summary: Enriches the input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummaryDocumentInputModel'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryDocumentOutputModel'
  /api/documents/voided-document/create:
    post:
      tags:
      - create
      summary: Create a XML file from the input
      parameters:
      - name: x-openubl-certificatekey
        in: header
        schema:
          type: string
      - name: x-openubl-privatekey
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoidedDocumentInputModel'
      responses:
        "200":
          description: OK
          content:
            text/xml:
              schema:
                type: string
  /api/documents/voided-document/enrich:
    post:
      tags:
      - enrich
      summary: Enriches the input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoidedDocumentInputModel'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoidedDocumentOutputModel'
  /api/documents/perception/create:
    post:
      parameters:
      - name: x-openubl-certificatekey
        in: header
        schema:
          type: string
      - name: x-openubl-privatekey
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerceptionInputModel'
      responses:
        "200":
          description: OK
  /api/documents/perception/enrich:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerceptionInputModel'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerceptionOutputModel'
  /api/documents/retention/create:
    post:
      parameters:
      - name: x-openubl-certificatekey
        in: header
        schema:
          type: string
      - name: x-openubl-privatekey
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetentionInputModel'
      responses:
        "200":
          description: OK
  /api/documents/retention/enrich:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetentionInputModel'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetentionOutputModel'
...
components:
  schemas:
...
    ListDocumentLineInputModel:
      type: array
      items:
        $ref: '#/components/schemas/DocumentLineInputModel'
...
    ListDocumentLineOutputModel:
      type: array
      items:
        $ref: '#/components/schemas/DocumentLineOutputModel'
...
    CreditNoteOutputModel:
      required:
      - cliente
      - firmante
      - impuestos
      - proveedor
      - totales
      - tipoDocumentoComprobanteAfectado
      - tipoNota
      type: object
      properties:
        cliente:
          allOf:
          - $ref: '#/components/schemas/ClienteOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/ClienteOutputModel'
          - nullable: false
        detalle:
          allOf:
          - $ref: '#/components/schemas/ListDocumentLineOutputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/DocumentLineOutputModel'
          - $ref: '#/components/schemas/ListDocumentLineOutputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/DocumentLineOutputModel'
        fechaEmision:
          pattern: \S
          type: string
          nullable: false
        firmante:
          allOf:
          - $ref: '#/components/schemas/FirmanteOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/FirmanteOutputModel'
          - nullable: false
        horaEmision:
          pattern: \S
          type: string
          nullable: false
        impuestos:
          allOf:
          - $ref: '#/components/schemas/DocumentImpuestosOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/DocumentImpuestosOutputModel'
          - nullable: false
        moneda:
          maxLength: 3
          minLength: 3
          pattern: \S
          type: string
          nullable: false
        proveedor:
          allOf:
          - $ref: '#/components/schemas/ProveedorOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/ProveedorOutputModel'
          - nullable: false
        serieNumero:
          pattern: \S
          type: string
          nullable: false
        totales:
          allOf:
          - $ref: '#/components/schemas/DocumentMonetaryTotalOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/DocumentMonetaryTotalOutputModel'
          - nullable: false
        descripcionSustentoDeNota:
          pattern: \S
          type: string
          nullable: false
        serieNumeroComprobanteAfectado:
          pattern: \S
          type: string
          nullable: false
        tipoDocumentoComprobanteAfectado:
          allOf:
          - $ref: '#/components/schemas/Catalog1'
          - enum:
            - BOLETA
            - COMPROBANTE_PAGO_SEAE
            - DOCUMENTOS_AFP
            - DOCUMENTOS_FINANCIEROS
            - FACTURA
            - GUIA_REMISION_REMITENTE
            - GUIA_REMISION_REMITENTE_COMPLEMENTARIA
            - GUIA_REMISION_TRANSPORTISTA
            - GUIA_REMISION_TRANSPORTISTA_COMPLEMENTARIA
            - NOTA_CREDITO
            - NOTA_DEBITO
            - PERCEPCION
            - PERCEPCION_VENTA_INTERNA
            - RETENCION
            - TICKET_MAQUINA_REGISTRADORA
            nullable: false
          - $ref: '#/components/schemas/Catalog1'
          - enum:
            - BOLETA
            - COMPROBANTE_PAGO_SEAE
            - DOCUMENTOS_AFP
            - DOCUMENTOS_FINANCIEROS
            - FACTURA
            - GUIA_REMISION_REMITENTE
            - GUIA_REMISION_REMITENTE_COMPLEMENTARIA
            - GUIA_REMISION_TRANSPORTISTA
            - GUIA_REMISION_TRANSPORTISTA_COMPLEMENTARIA
            - NOTA_CREDITO
            - NOTA_DEBITO
            - PERCEPCION
            - PERCEPCION_VENTA_INTERNA
            - RETENCION
            - TICKET_MAQUINA_REGISTRADORA
            nullable: false
        tipoNota:
          allOf:
          - $ref: '#/components/schemas/Catalog9'
          - enum:
            - ANULACION_DE_LA_OPERACION
            - ANULACION_POR_ERROR_EN_EL_RUC
            - BONIFICACION
            - CORRECCION_POR_ERROR_EN_LA_DESCRIPCION
            - DESCUENTO_GLOBAL
            - DESCUENTO_POR_ITEM
            - DEVOLUCION_POR_ITEM
            - DEVOLUCION_TOTAL
            - DISMINUCION_EN_EL_VALOR
            - OTROS_CONCEPTOS
            nullable: false
          - $ref: '#/components/schemas/Catalog9'
          - enum:
            - ANULACION_DE_LA_OPERACION
            - ANULACION_POR_ERROR_EN_EL_RUC
            - BONIFICACION
            - CORRECCION_POR_ERROR_EN_LA_DESCRIPCION
            - DESCUENTO_GLOBAL
            - DESCUENTO_POR_ITEM
            - DEVOLUCION_POR_ITEM
            - DEVOLUCION_TOTAL
            - DISMINUCION_EN_EL_VALOR
            - OTROS_CONCEPTOS
            nullable: false
...
    DebitNoteOutputModel:
      required:
      - cliente
      - firmante
      - impuestos
      - proveedor
      - totales
      - tipoDocumentoComprobanteAfectado
      - tipoNota
      type: object
      properties:
        cliente:
          allOf:
          - $ref: '#/components/schemas/ClienteOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/ClienteOutputModel'
          - nullable: false
        detalle:
          allOf:
          - $ref: '#/components/schemas/ListDocumentLineOutputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/DocumentLineOutputModel'
          - $ref: '#/components/schemas/ListDocumentLineOutputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/DocumentLineOutputModel'
        fechaEmision:
          pattern: \S
          type: string
          nullable: false
        firmante:
          allOf:
          - $ref: '#/components/schemas/FirmanteOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/FirmanteOutputModel'
          - nullable: false
        horaEmision:
          pattern: \S
          type: string
          nullable: false
        impuestos:
          allOf:
          - $ref: '#/components/schemas/DocumentImpuestosOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/DocumentImpuestosOutputModel'
          - nullable: false
        moneda:
          maxLength: 3
          minLength: 3
          pattern: \S
          type: string
          nullable: false
        proveedor:
          allOf:
          - $ref: '#/components/schemas/ProveedorOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/ProveedorOutputModel'
          - nullable: false
        serieNumero:
          pattern: \S
          type: string
          nullable: false
        totales:
          allOf:
          - $ref: '#/components/schemas/DocumentMonetaryTotalOutputModel'
          - nullable: false
          - $ref: '#/components/schemas/DocumentMonetaryTotalOutputModel'
          - nullable: false
        descripcionSustentoDeNota:
          pattern: \S
          type: string
          nullable: false
        serieNumeroComprobanteAfectado:
          pattern: \S
          type: string
          nullable: false
        tipoDocumentoComprobanteAfectado:
          allOf:
          - $ref: '#/components/schemas/Catalog1'
          - enum:
            - BOLETA
            - COMPROBANTE_PAGO_SEAE
            - DOCUMENTOS_AFP
            - DOCUMENTOS_FINANCIEROS
            - FACTURA
            - GUIA_REMISION_REMITENTE
            - GUIA_REMISION_REMITENTE_COMPLEMENTARIA
            - GUIA_REMISION_TRANSPORTISTA
            - GUIA_REMISION_TRANSPORTISTA_COMPLEMENTARIA
            - NOTA_CREDITO
            - NOTA_DEBITO
            - PERCEPCION
            - PERCEPCION_VENTA_INTERNA
            - RETENCION
            - TICKET_MAQUINA_REGISTRADORA
            nullable: false
          - $ref: '#/components/schemas/Catalog1'
          - enum:
            - BOLETA
            - COMPROBANTE_PAGO_SEAE
            - DOCUMENTOS_AFP
            - DOCUMENTOS_FINANCIEROS
            - FACTURA
            - GUIA_REMISION_REMITENTE
            - GUIA_REMISION_REMITENTE_COMPLEMENTARIA
            - GUIA_REMISION_TRANSPORTISTA
            - GUIA_REMISION_TRANSPORTISTA_COMPLEMENTARIA
            - NOTA_CREDITO
            - NOTA_DEBITO
            - PERCEPCION
            - PERCEPCION_VENTA_INTERNA
            - RETENCION
            - TICKET_MAQUINA_REGISTRADORA
            nullable: false
        tipoNota:
          allOf:
          - $ref: '#/components/schemas/Catalog10'
          - enum:
            - AUMENTO_EN_EL_VALOR
            - INTERES_POR_MORA
            - PENALIDAD_OTROS_CONCEPTOS
            nullable: false
          - $ref: '#/components/schemas/Catalog10'
          - enum:
            - AUMENTO_EN_EL_VALOR
            - INTERES_POR_MORA
            - PENALIDAD_OTROS_CONCEPTOS
            nullable: false
...
    CreditNoteInputModel:
      required:
      - cliente
      - detalle
      - numero
      - proveedor
      - serie
      - descripcionSustentoDeNota
      - serieNumeroComprobanteAfectado
      type: object
      properties:
        cliente:
          allOf:
          - $ref: '#/components/schemas/ClienteInputModel'
          - nullable: false
        detalle:
          allOf:
          - $ref: '#/components/schemas/ListDocumentLineInputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/DocumentLineInputModel'
            nullable: false
        fechaEmision:
          format: int64
          type: integer
        firmante:
          $ref: '#/components/schemas/FirmanteInputModel'
        numero:
          format: int32
          maximum: 99999999
          minimum: 1
          type: integer
          nullable: false
        proveedor:
          allOf:
          - $ref: '#/components/schemas/ProveedorInputModel'
          - nullable: false
        serie:
          maxLength: 4
          minLength: 4
          pattern: \S
          type: string
          nullable: false
        descripcionSustentoDeNota:
          pattern: \S
          type: string
          nullable: false
        serieNumeroComprobanteAfectado:
          pattern: \S
          type: string
          nullable: false
        tipoNota:
          type: string
    DebitNoteInputModel:
      required:
      - cliente
      - detalle
      - numero
      - proveedor
      - serie
      - descripcionSustentoDeNota
      - serieNumeroComprobanteAfectado
      type: object
      properties:
        cliente:
          allOf:
          - $ref: '#/components/schemas/ClienteInputModel'
          - nullable: false
        detalle:
          allOf:
          - $ref: '#/components/schemas/ListDocumentLineInputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/DocumentLineInputModel'
            nullable: false
        fechaEmision:
          format: int64
          type: integer
        firmante:
          $ref: '#/components/schemas/FirmanteInputModel'
        numero:
          format: int32
          maximum: 99999999
          minimum: 1
          type: integer
          nullable: false
        proveedor:
          allOf:
          - $ref: '#/components/schemas/ProveedorInputModel'
          - nullable: false
        serie:
          maxLength: 4
          minLength: 4
          pattern: \S
          type: string
          nullable: false
        descripcionSustentoDeNota:
          pattern: \S
          type: string
          nullable: false
        serieNumeroComprobanteAfectado:
          pattern: \S
          type: string
          nullable: false
        tipoNota:
          type: string
    InvoiceInputModel:
      required:
      - cliente
      - detalle
      - numero
      - proveedor
      - serie
      type: object
      properties:
        cliente:
          allOf:
          - $ref: '#/components/schemas/ClienteInputModel'
          - nullable: false
        detalle:
          allOf:
          - $ref: '#/components/schemas/ListDocumentLineInputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/DocumentLineInputModel'
            nullable: false
        fechaEmision:
          format: int64
          type: integer
        firmante:
          $ref: '#/components/schemas/FirmanteInputModel'
        numero:
          format: int32
          maximum: 99999999
          minimum: 1
          type: integer
          nullable: false
        proveedor:
          allOf:
          - $ref: '#/components/schemas/ProveedorInputModel'
          - nullable: false
        serie:
          maxLength: 4
          minLength: 4
          pattern: \S
          type: string
          nullable: false
...
    SummaryDocumentInputModel:
      required:
      - detalle
      - fechaEmisionDeComprobantesAsociados
      - numero
      - proveedor
      type: object
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/ListSummaryDocumentLineInputModel'
          - minItems: 1
            items:
              $ref: '#/components/schemas/SummaryDocumentLineInputModel'
            nullable: false
        fechaEmision:
          format: int64
          type: integer
        fechaEmisionDeComprobantesAsociados:
          format: int64
          type: integer
          nullable: false
        firmante:
          $ref: '#/components/schemas/FirmanteInputModel'
        numero:
          format: int32
          minimum: 1
          type: integer
          nullable: false
        proveedor:
          allOf:
          - $ref: '#/components/schemas/ProveedorInputModel'
          - nullable: false
...